Add a test showing inlay hints behaviour for un-named parameters
Summary:
In the case of un-named parameters, we produce fake names of the form `ArgN`. While this is useful in the context of *signature help* or when adding an EDoc stub, it does not make sense to produce inlay hints with generated names since they add unnecessary noise. Therefore, we need a way to distinguish a real name from a generated one.

This diff simply introduces a test case showing the current behaviour, which will be modified later in the stack.

Reviewed By: alanz

Differential Revision: D47985068

fbshipit-source-id: 6fdb4556a525ede04ebb8c5b8e7ae8d91689a00d