Do not produce inlay hints for un-named parameters
Summary: In the case of un-named parameters, we produce fake names of the form `ArgN`. While this is useful in the context of *signature help* or when adding an EDoc stub, it does not make sense to produce inlay hints with generated names since they add unnecessary noise. Therefore, we need a way to distinguish a real name from a generated one.

Reviewed By: alanz

Differential Revision: D47985056

fbshipit-source-id: 89b2e9dbece0a2190536c99f0a269dca4115fd93