Do not require single file fixtures for inlay hints tests
Summary: For inlay hints, we need to test more complex scenarios involving multiple files. Therefore, tweak the testing framework to allow multiple fixtures. The `~` marker represents the current cursor position, which here we use to identify the currently active buffer (the one we extract annotations from). As a follow-up, we could not require the marker for single fixtures.

Reviewed By: alanz

Differential Revision: D47952713

fbshipit-source-id: 2d3b4ffb87d60f5a8df8c1b4e31218339846d200