Exclude ad-hoc diagnostics from those recursively applied
Summary:
# Context
Ad-hoc diagnostics are usually used for one-off codemodding and the `--recursive` options will apply other diagnostics/rewrite-rules until a fixpoint is found, in order to do any post-codemod cleanup.

# The problem
Say we have an ad-hoc diagnostic and we are using `--line-from`/`--line-to` to limit it's application to a particular line/range, etc. If we want to use `--recursive` on top, then the ad-hoc diagnostic will end up applied outside of the given range as well/

# This diff
As a quick mitigation, we exclude ad-hoc diagnostics from the set applied recursively until a more general solution is implemented (e.g. controlling the set of recursive diagnostics via cli args?).

Alternatively, we could have excluded the given diagnostic from the recursive set, but we still have the issue that we sometimes want to try to apply several ad-hoc diagnostics and find out which lead to changes (e.g. for user reporting), so excluding only the given one would not be enough (as the rest would also be applied)

Reviewed By: alanz

Differential Revision: D48030625

fbshipit-source-id: e61b48916c74a4268f1946859914d0bb04cdec4f