Introduce diagnostic categories
Summary: At the moment, diagnostics can be marked with an "experimental" flag, that controls whether they are applied or not. We generalize this to a notion of "diagnostic category" and will allow diagnostics to have zero-or-more of this. While there is only one possible category at the moment, we introduce another one in the next diff introduces another one.

Reviewed By: alanz

Differential Revision: D48069474

fbshipit-source-id: e1ae716603506b5fc5971bc675111abe41e53dd4