Categorize diagnostics used by lint_cli as SimplificationRules
Summary:
# Context
Currently, when running `elp lint --recursive`, after applying the first rule, a fixpoing of all non-adhoc diagnostics is applied, with the intention of cleaning-up the code after the initial transformation

# Problem
We had very little control of which rules participated in this fixpoint. We started to filter-out ad-hoc rules as a hack, but this doesn't scale well, as during codemods we may want to introduce additional ad-hoc rules to be used only for simplification.

# This diff
We levarage diagnostic categories and introduce a category of "simplification rules", and let lint_cli use only those during the fixpoint.

Reviewed By: alanz

Differential Revision: D48069473

fbshipit-source-id: e2533e017772326b81451846f69df32f0f3d2710