Dependent Header Diagnostics
Summary:
### Overview

Header files are processed lazily by the Erlang compiler. If a header file contains a macro which refers to a record not defined within the same file, `erlc` won't complain. But this represents a bad smell and ELP can detect it.

This diff introduces the ability to identify header files which are not "self contained". For the time being, the analysis is limited to records only.

### To Test This

Run the unit tests for the introduced module
```
cargo test dependent_header
```

Run the linter for a single diagnostic
```
elp lint --project [PATH_TO_PROJECT] --diagnostic-filter W0015
```

Notice that, since currently header files are not processed by the `elp lint` command, the above still returns an empty result so it's not valid as a test.

Reviewed By: alanz

Differential Revision: D47366856

fbshipit-source-id: 2094e9c1cb7035ac8d7a189387897ca6502955d8