do not generate diagnostics for files not part of the project model
Summary:
If a file is not known to the project model, it appears in the catch-all source root.
Store the `SourceRootId` of this, and when classifying files, used to decide whether to send diagnostics or not, mark them as `FileKind::OutsideProjectModel` so we do not process them.

In future, when we deal with out-of-project erlang files we can revisit this.

It does mean we will not produce basic diagnostics for random .erl files we open.

To obviate that, we could add a REGEXP path check in the `file_kind` function, to exclude blacklisted directories but otherwise process .erl files.

Reviewed By: perehonchuk

Differential Revision: D48313381

fbshipit-source-id: 8e892e21d1c9c309593aaff8cf43006fe0ca1ef9