check regexp of paths for ones not in project model
Summary:
When classifying a FileKind for a given file, if it is outside the existing project model check if it is on a list of blacklisted directories, otherwise process it.

This means we do get basic diagnostics on files not in the existing project model, and not on the blacklist.

This is a follow-up to D48313381

Reviewed By: perehonchuk

Differential Revision: D48317712

fbshipit-source-id: 36660a75d3c70a2077c813a7c43edeb409caee47