BE: make ancestors_at_offset panic-safe
Summary:
We are getting occasional panics from invoking `ancestors_at_offset` where the offset is invalid.
This is most likely during completion calculation, where we are working with out-of-sync code.
Harden against this, by moving the check done by the failing assertion into `ancestors_with_offset`, and return a `None` instead of panicking.

Reviewed By: ir-regular

Differential Revision: D48430278

fbshipit-source-id: 1afbeaaf194078c3358920f803b0efd451e1f778