group related diagnostics
Summary:
When we have a syntax error in a function, the erlang_service diagnostics report undefined function errors wherever it is referenced.
Group them, by checking for an ELP-reported syntax error that occurs in a function, and then add any erlang_service undefined diagnostics to the related info of the syntax error.

We do it at the point we gather diagnostics to be sent to the client, to obviate any timing issues between the erlang_service and native diagnostics arriving.

{F1066114034}

Reviewed By: robertoaloi

Differential Revision: D48389172

fbshipit-source-id: fa4011f3188e79efff80c765ca5ea29e72447323