BE: Introduce AstClauseId
Summary:
When we lower an AST to HIR, we end up with clauses stored in an arena accessed by `ClauseIdx`.
The intention is that these are opaque, and you can only be given a valid one.
When we move from AST level processing to HIR level, we construct a `ClauseIdx` by counting the clauses in the AST to the current location.

In practice, we are seeing panic's where this count is invalid, presumably for completion on an ill-formed clause that we are unable to lower.

Explicitly tag the derived `ClauseId`s as `AstClauseId`, and require a validation step before they can be used.

Reviewed By: ir-regular

Differential Revision: D48437075

fbshipit-source-id: f026c5be8d38ab104506c7bac4a09c21d4fdac0d