Add fold::Strategy::SurfaceOnly for fold_expr
Summary:
`hir::fold` allows traverals over the HIR AST.
When you call it, you provide a `Body` which can be indexed either to ignore `Expr::MacroCall`, or not.
The `Expr::MacroCall` has inside it an `ExprId` for the full expansion, as well as a `Vec<ExprId>` for the specific macro arguments.

Provide a new enum for `hir::Strategy`, `SurfaceOnly`, so that when called with a macro-aware Body, the traversal will not descend into the macro expansion, but will call back for the arguments instead.

This implements it for the expr callback only

Reviewed By: perehonchuk

Differential Revision: D48606794

fbshipit-source-id: d2c3bac35c4ce2a28446a65f045a8ac9fa5b61ff