Add config file
Summary:
Add `--read-config` option to `elp lint`.
This looks for a configuration file called `.elp_lint.toml` in the project directory used for the command.
If it exists, it is read and the sets of enabled and disabled lint codes applied.
This makes passing `--diagnostic-filter` optional, and also means that the enabled lints will have their fixes applied if the command is used this way.

In general, the intent is that the config file is used when using `elp lint` as a linter for CI, to report issues on a given code base.

Reviewed By: ir-regular

Differential Revision: D48648912

fbshipit-source-id: e6f2e3e8b545254bc4f63c493cc21d03beb71b57