Diagnostic for deprecated function usage
Summary:
This change introduces a new diagnostic (code `W0016`) to mark deprecated diagnostics, effectively superseding the `deprecated_function_calls` analysis in XRef.

The diagnostic leverages the same `-deprecated` attribute that XRef uses, but it also displays the optional description (the third field in the attribute) at the call-site, making it easier to guide developers hovering a deprecated function call (the description is unused by XRef itself).

There is currently no support for the `-ignore_xref` attribute used by rebar3 to silent specific instances of this diagnostic. Instead, the regular `% elp:ignore` comments can be used.

The deprecation information can be augmented with extra details such as:

* A URI to a page containing additional information about the deprecation
* A severity for the deprecation (info, warning, error)

Later on this can be expanded with a DSL-like syntax that code-mods (i.e. autofixes) can leverage (e.g. by removing a no-op function or by replacing a pattern with something else). For the time being the additional information is simply specified in code.

Architectural note: it looks to me like this new diagnostic and the `application_env` and `cross_node_eval` ones are essentially the same thing. With a little refactoring we should be able to merge them into one.

Reviewed By: alanz

Differential Revision: D48516481

fbshipit-source-id: c03c07a5307429a2559441c1a07078389aa4eb27