Add GitHub workflows to build/upload the VS Code extension
Summary:
This represents the first step towards automating building and publishing of the ELP VS Code extension.
At this stage the extension does not include the ELP binary yet (the server needs to manually be installed and the extension can be pointed to it via a setting). This will be changed later.

The version of the extension has been changed to match the version of the next ELP release.

The pipeline contain lot of duplication. That can be fixed as a follow-up.

Reviewed By: alanz

Differential Revision: D48831779

fbshipit-source-id: c93382a27afc51ae2331a9452de71df054e00f1c