Expand ?FUNCTION_NAME with args
Summary:
The `?FUNCTION_NAME` macro resolves to an atom, being the name of the function it is in.
But if it is called with arguments it is still a valid expansion, just a call to a function with the same base name as the current one.
Update macro expansion and lowering for this.

Reviewed By: robertoaloi

Differential Revision: D49012586

fbshipit-source-id: bbb2fb3e5aff5648f3c108ffb716b0fa22b755d5