improve missing X diagnostic reported location
Summary:
The tree-sitter grammar is set up to have optional semicolons between function clauses, and commas in lists, so we can still process the AST if they are missing.
We check after the fact if they are missing and report diagnostics for them.
Rather than attach the diagnostic to the entire clause with a missing semicolon, attach it to the last non-whitespace item in the clause.

Reviewed By: robertoaloi

Differential Revision: D49094298

fbshipit-source-id: 2d1d735cc38278f3528bd52aa99d8351166aa199