Publish platform specific extension, instead of rebuilding it
Summary:
The `vsce publish` command, when not given a `--packagePath` argument, will re-package the extension instead of using the already existing artifact. The new package will then be published as "universal" since the `--target` argument is not given to the `publish` command. This change fixes it.

Reference: https://code.visualstudio.com/api/working-with-extensions/publishing-extension

Reviewed By: alanz

Differential Revision: D49152745

fbshipit-source-id: 41b17f67916eba4658ff0894349a040ab4f6b6af