show diagnostic code in CI from ELP lints
Summary:
We generate lint warnings in CI via `elp lint`, and there is a mechanism to ignore them using `elp:ignore` comments.
Provide both the code and the label for this use case.
Note: `elp:ignore` will recognise either the code or the label. But since the error index works against the code, we show both.

Reviewed By: robertoaloi

Differential Revision: D49185528

fbshipit-source-id: 9a55fb22f801cb8fd199b8efa87e37a1e28ce1b1