replace_call: use resolved function module
Summary:
The `codemod_helpers::replace_call` provides a callback to check that the replacement should go ahead for a specific matched `MFA`.
Rather than using the match spec, which may have an inaccurate module name for local or imported functions, resolve the call target and generate the `MFA` from that.

Reviewed By: jcpetruzza

Differential Revision: D49456804

fbshipit-source-id: 8cc6cd2fcce1e32c2496d34f9246f99144a241bc