Add a map from DefineId to FormId in FormList
Summary:
Except for `PPDirective`, for each `FormList` item, we can create a `FormIdx` by simply wrapping it.
The `PPDirective` has an enum subtype which is returned when iterating over individual directives.
Maintain a map from this sub-indexes to the appropriate `FormIdx`, so we can easily find the host form.

This is required for `fold_file`, where we need a `FormIdx` in the context, so the callback can take appropriate action.

Reviewed By: perehonchuk

Differential Revision: D49321952

fbshipit-source-id: 7cf4e7dd476eec7736fe21306562fa9f1efde552