use single callback for fold
Summary:
We currently have distinct callbacks for expr, pat and term.
As we extend to folding over the full file, this will need to extend to types too.
Rather then letting callbacks proliferate, reduce it to a single one, taking an `AnyExprId` and `AnyExpr`, which are enumerations over expr, pat etc.

This is a big diff, but most of it is just mechanical replacement.

Reviewed By: perehonchuk, TD5

Differential Revision: D49322286

fbshipit-source-id: fff588767738603e0f9f542346aef410c89585ce