tree-sitter conversion. Capture top level ERROR structure
Summary:
Tree-sitter can return an AST where the top level is an ERROR node, because one of the forms would not parse.  But the rest of the forms do parse correctly, the parser is just not able to do the final `reduce` operation to generate a form list.
When this happens, the incomplete results are included as children of the ERROR node. Adjust the tree-sitter to rowan AST conversion to capture this detailed structure too, rather than throwing it away.

Reviewed By: perehonchuk

Differential Revision: D49462782

fbshipit-source-id: 99d6aaacc4f61cd7937fddc60a344c21cf683449