Introduce HirIdx, a file-scope AnyExprId index
Summary:
An `AnyExprId` is only valid in the `Body` it indexes into.
Introduce `HirIdx` that combines it with a `FormIdx` to identify the body being used.

This allows us to uniquely index an HIR AST item in a file.

Update the `fold::AnyCallBackCtx` to use this instead of `AnyExprId`.

This paves the way for a full-file fold.

Reviewed By: robertoaloi

Differential Revision: D49682989

fbshipit-source-id: 415712b0e9e28d7924e4b6c78f7eaa10e62d39c5