Report EDoc scanning issues
Summary:
Store scanning errors in the process dictionary, so they can be picked by the `edoc_report` module.

Also raise the severity for EDoc diagnostics back to warning. The issue we were trying to solve does not seem to occur any longer, so it makes sense for EDoc problems to be more visible.

Reviewed By: alanz

Differential Revision: D49729116

fbshipit-source-id: c24678ac1362842178d548b75e7170f43195c69f