suite/0 codemod helpers
Summary:
Add a helper function to ensure a given `{key, value}` tuple exists in the `suite/0` function of a test suite.
If the function does not exist, it is created.
The function is exported.
If the `key` is already in the `suite/0` list its value is updated, otherwise it is added to the end.

Reviewed By: jcpetruzza

Differential Revision: D49643856

fbshipit-source-id: b0bd6fd6d77abf7a25706a74b082aa89ea422da0