Introduce start of fold_file function
Summary:
This is the start of a diff stack.
It introduces the `fold_file` function which does a fold over every form in a file, with a callback for each visited element.
The callback includes the `FormId`, so the details of the relevent form can be accessed in the callback.

Reviewed By: perehonchuk

Differential Revision: D49777940

fbshipit-source-id: b5a317f95a7b8153732ba3cc1baae0743aa2149a