BE: Remove replace_call::Replacement::Compo
Summary:
The processing is not safe.
As defined, for the first changeit analyzes the code, generates an edit, and stores that edit.
Without applying the first edit, it analyzes the original code, and generates another edit.
The first edit should be applied to the source first, and then the second change evaluated.

This is done in `elp_lint.rs`, which makes sure to apply a single change, and re-analyzes the code before checking for another one.

Reviewed By: acw224

Differential Revision: D49861267

fbshipit-source-id: c770387f2c19f7aefb7e314430fa689a690ac7dd