Refactor types on hover logic
Summary:
Refactor types on hover to avoid unstable/dubious conversions. The markdown documentation, including links to references, is now generated using `snapshot` and `to_proto` capabilities, to stay as close to the IDE as possible. To that end:
- Functions in `eqwalizer` do not generate documentation anymore, the module provides a function returning type references of the form `Vec<(SmolStr, FileRange)>` instead.
- Doc generation is moved to `handlers::handle_hover`, which fetches type and references from `eqwalizer`, and then uses `to_proto` to generate markdown links.

This is much more robust and solves the issue of links being broken on ODs

Reviewed By: alanz

Differential Revision: D49865396

fbshipit-source-id: 669102bd0cbab2af7ca1705987035ecfbb1591c2