Add FormList::get to map FormIdx to Form
Summary:
This makes it easier to `match` on the result of the `FormList::get` call for a `FormIdx`, rather than dereferencing for each possible `FormIdx` in the `match`.

We can't create an `Index` instance for this because the return type needs a reference, and we must wrap the inner dereference.

Reviewed By: perehonchuk

Differential Revision: D49830564

fbshipit-source-id: be009131612e247b37d10809e6de6b3fdd272543