Avoid redundant edits when adding to suite/0
Summary: We plan to use add_to_suite_0() in a codemod/diagnostic, but currently when the thing we are adding to the suite is already there, we still generate an edit (that will produce no visible change). Since we will take into account whether there are any edits to decide whether to produce the diagnostic, we want to avoid this type of redundant edits.

Reviewed By: alanz

Differential Revision: D49910211

fbshipit-source-id: 2f28cdadaa1190b213ecb06f2f22816cdfdbf0b9