replace call: provide extra_info for the fix too
Summary:
`find_call_in_function` takes a callback to check if a given call needs a diagnostic, which returned a string to be included in the diagnostic.
Extend this to return another string too, to be used for any fix attached to the diagnostic.

This lets us have a fix like "replace foo:bar() with ok", and becomes meaningful when these are exposed in the IDE too.

Reviewed By: ir-regular

Differential Revision: D49952485

fbshipit-source-id: 8263d04999d98cc8ea36ca7f242bfa00e72b2c68