read config file to IDE
Summary:
If the `experimental` flag is set, read ad hoc lints from the `.elp_lint.toml` file.
It does not use the `enabled` and `disabled` lints though.

It looks for the config file based on the project root of the first project file in the server state.
In future it should perhaps keep per-project settings.

This is intended as a mechanism to easily debug potential lints being authored, to understand easily their effect on live code.

Reviewed By: ir-regular

Differential Revision: D50010345

fbshipit-source-id: dba2702a8f0131f0c7b040a58d6b02e94f594064