Add diagnostic/codemod to onboard to auto_meckanic
Summary:
# Context
We have recently launched support for automatically dealing with clean-up of meck mocks ([details](https://fb.workplace.com/groups/2543415995888283/permalink/3754089388154265/)).

# Problem
Because enabling this globally breaks lots of suites, tests need to opt-in at the moment. The onboarding mechanism is mostly mechanic:
  - add an element to the suite/0 proplist
  - remove all meck cleanups from end_per_*
  - some init_per_* functions contain defensive meck cleanups, remove these as well
  - many testcases contain trailing calls to meck:unload, etc, meant to be manual cleanups, remove these and any initial defensive onces. The ones in the middle need to be preserved since they can be part of the test logic

# This diff
We add a diagnostic to ELP that marks when a suite can be onboarded to auto_meckanic and performs all mechanic transformations.

We plan to integrate this diagnostic with the CodemodService so that we can automate the onboarding of all suites that don't need manual intervention

Reviewed By: TheGeorge

Differential Revision: D49841674

fbshipit-source-id: 908697aefe8bcca4163bb547a6d8a29172970c13