Add TextRange to LabeledDiagnostic
Summary:
Diff 2 of 3

The `RangedSet` of enclosing `Form` ranges from ELP native syntax errors is based on a `TextRange`.
When we do the combining, the diagnostics have been converted to LSP ones, using LSP `Range` types.
So keep the original `TextRange` with each diagnostic.

These will be used in the final diff for filtering out `erlang_services` in forms that already have ELP native diagnostics

Reviewed By: TD5

Differential Revision: D50222467

fbshipit-source-id: 5c202c2c347f889067d5cc173570035d9e91b645