Global file path resolution for type references
Summary:
When resolving the path to a type declaration, searching the source root of the current file is not enough, as the declaration can take place in another app root.

Since we do not have access to the VFS at this point, perform a simple lookup of all source roots and pick the first that resolves successfully.

Performance does not seem to be an issue, but the operation can be cached if it becomes one.

Reviewed By: alanz

Differential Revision: D50223628

fbshipit-source-id: 2283dd2b3bf0291398cd669a75f26b0f49dbc343