Clamp client positions and ranges to fall within referenced file
Summary:
We see a number of panics in our logs from a client-provided range or position no longer corresponding to a location in our VFS version of the file.
Introduce functions to clamp any range or offset provided to the limits of the file being processed, and apply this in any server handler receiving a range or position.

This should get rid of the last of the panics from out of range offsets

Reviewed By: ir-regular

Differential Revision: D50267952

fbshipit-source-id: 2613f7fb3d4acafcd6500e749ea5aee8b4abff62