add json project manifest
Summary:
re-implementation from the mini-elp.
the idea is the same, we have `build_info.json` file with the same structure like rebar's build info. The only difference that mini-elp represents this manifest as fake rebar `ProjectBuildData`. My approach is to have separate `ProjectBuildData::Static`  which we will use later for `NoManifest` project loading

Reviewed By: alanz

Differential Revision: D49692668

fbshipit-source-id: 6718c78939a56e9b415b2c2b527fd0f047a1cdda