make eqwalizer parse-all results deterministic on parse failure
Summary:
Eqwalizer can only generate a result when it is given files without errors.
If it is unable to compile a particulare file, it returns a status of `NoAst` with the module name.
In a test, we check for this behaviour while processing a whole project, but also look for errors reported on a specific module.
The test project has more than one that can fail.
In the case of more than one module failure, make the combination process deterministic so the test will always report detailed errors on the same module, however the parallel processing does the actual eqwalizer process.

Reviewed By: michalmuskala

Differential Revision: D50557011

fbshipit-source-id: fb5eb6e7c29d2a976d159f3a772222390d957edc