Manage change ranges when recursing
Summary:
For `elp lint --recursive` we follow a strategy of applying one lint at a time, and then re-parsing the changed file.
We then apply clean-up fixes, but only for the locations that had changed.
We choose the enclosing form as the unit of granularity for this "change" check.
Keep a global list of changes, so that clean-up happens to all requisite changed forms, even if they were changed in an ealier cycle.

The recursive lint test is updated to demonstrate this, we have two locations that require fixes and cleanups, and two pre-existing issues that could be cleaned up. They remain untouched, but both required changes and cleanups happen

Reviewed By: jcpetruzza

Differential Revision: D50653839

fbshipit-source-id: c4f9f80a7ff1a971006d2b6bb2d7531dc0e17fb4