Evaluate `all/0` and `groups/0` CT Callbacks via Erlang Service
Summary:
Until now we have been parsing the content of the `all/0` and `groups/0` functions from Common Test test suites. This information was used to produce the "Debug" and "Run" code lenses for CT testcases and the "unreachable testcase" diagnostic.

This approach had a significant limitation, though, since it only worked for cases where the `all/0` and `groups/0` functions contained literals only. This limitation is solved by this change, which evaluates the `all/0` and `groups/0` functions via the Erlang Service.

This change also sets the ground to compute more granular Common Test diagnostics (e.g. _malformed test definition_) and to provide insights about a test suite which are useful for linters and code-mods (e.g. the content of the `suite/0` function).

**Note about the tests:** to be able to test features such as runnables and CT diagnostics programmatically, I had to apply two changes to the testing framework:

* Ensure the Erlang Service is started when testing runnables and CT diagnostics
* Introduce a new piece of metadata (`scratch_buffer:true`) that can be added to a test fixture to inform the framework to dump a copy of the fixture to the file system, so that the Erlang Service can compute CT info correctly

Reviewed By: michalmuskala

Differential Revision: D50691832

fbshipit-source-id: 628b4cd24807813140363a4f7670b15433684c83