Remove SimplificationRule category from deprecated_functions diagnostic
Summary:
# Context
Some diagnostics double as "simplification rules" that are applied recursively if `elp lint --apply-fix --recursive` is used and take care of doing clean-ups that are sound. This let us write complex codemods in a modular fashion.

# Problem
The fix for a deprecated function is to add an ignore_xref() annotation. This is definitely not sound and most codemods are unlikely to introduce the usage of deprecated functions, so this is not a clean-up that we want. As it is, codemodding a file that contains deprecated functions, will end up adding several ignore_xref() annotations.

# This diff
This is likely a copy-and-paste issue, so just removing the SimplificationRule annotation

Reviewed By: michalmuskala, alanz

Differential Revision: D50791133

fbshipit-source-id: a4e6f4fae7e06e86f08a2f071575ac50f9ffe4e3