Add a salsa file revision, to manage files on disk
Summary:
We have situations where certain diagnostics are not always updated on save.
These are the ones calculated by an external process needing to read the contents of a file on disk.

# Problem

The LSP client sends a `DidChangeTextDocument` message when a file changes, and we generate native diagnostics from the changed file contents, and these calculations are cached.
The LSP client also sends a `DidSaveTextDocument` to indicate a document has been saved to disk, and we use that to trigger the diagnostics that read the file on disk.
BUT, for some reason we sometimes call `module_ast` which reads the contents of a file on disk after a `DidChangeTextDocument`. This caches in salsa the unsaved state of the file, against the in-memory state.

When we call it again after actually writing the document to disk (in the client), salsa re-uses the prior value.

# Solution

We add a `file_revision` salsa database entry, initially set to zero for each file, and which is bumped by one on every change, and also on every save.
Any process requiring the on-disk version does a throwaway read of the database entry to register a dependency in salsa, and so it will be called when the version is bumped on the save too.

# Niggle

I was unable to trace the origin of the `module_ast` call after `DidSave`, as it will be working with out of date information.  But this fix restores coherence on save, so is useful in any case.

Reviewed By: robertoaloi, TD5

Differential Revision: D50880315

fbshipit-source-id: ffe8b32577f2545405e7f729afbcaac1e07f8bad