Undefined function diagnostic
Summary:
The L1227 diagnostic (part of the Erlang linter) can warn the user about local calls to undefined functions.
The new W0017 diagnostic does the same for fully qualified function calls, emulating the XRef "undefined_calls" analysis, but basing results on the internal ELP database.

While this cannot replace XRef for a production system due to ELP not being (yet) able to distinguish "flavours" of compilation (e.g. prod vs test), we can be sure that if a function is reported as undefined by ELP it can mean one of two thing:

* There's an actual bug in the code (function is really undefined)
* There's a bug in ELP (so it's good to surface it anyway)

The new diagnostic is marked as experimental for the time being.
Calls to external functions via `spawn()` or `apply()` are out of scope and can be added as a follow-up.

The changes to un-related tests are simply to exclude the new diagnostic from the config, which would result in spurious errors in those tests (since sometimes they contain references to undefined functions).

Reviewed By: alanz

Differential Revision: D50877872

fbshipit-source-id: f410189e8e9caf6f32d0961d4255f1b4f2ad76d6