capture changed forms after applying fix
Summary:
After applying a fix to a file we generate a diff to track the forms that changed.
The locations in the diff refer to the changed file, since that is what we are working with for the next fix.
Calculate which form is affected when we parse the changed source, rather than using the old.

Reviewed By: jcpetruzza

Differential Revision: D50846750

fbshipit-source-id: b80c4a986878b57a03c08c3e80488dbc4c32482b