Allow edits on previously edited ranges to be ignored
Summary:
# Context
We use the SourceChangeBuilder to build fixes for diagnostics and codemods, by appending insertions, delietions and replacements. At the moment, the ranges of these edits need to be disjoint, or the underlying TextEditBuilder will panic

# Problem
For complex codemods, we'd like to be able to allow some form of confluent edits. For example, let's say we are replacing all calls to foo:bar/2 by `ok`, and we get something like:

```
foo:bar(X, foo:bar(Y, Z))
```

The end result should be `ok` regardless which occurrences of `foo:bar/2` we replace by `ok` first. However at the moment there is no easy way to implement something like this, since whichever is the second occurrence we try to replace by `ok`, the operation will panic.

# This diff
We make it so that it is possible to implement the above safely using a TopDown fold strategy. The idea is to keep track of which ranges have already been modified and opt-in to not failing if an attempted edit is fully contained in a previously edited range. In the example above, this would mean that being top-down we will first replace `foo:bar(X, foo:bar(Y, Z))` by `ok`, and when later in the fold we try to replace `foo:bar(Y, Z)` by `ok`, this will be ignored since it is completely subsumed by later.

Reviewed By: TheGeorge

Differential Revision: D50983054

fbshipit-source-id: da9d48d3f581ec6fc9c4c613a95f11cc23457a3d