remove extraneous set_file_revision
Summary:
D50880315 brought in file revisions to ensure the salsa cache was invalidated when a file was saved, and we requested eqwalizer and erlang_service diagnostics.

It seems there is some kind of race condition on using `raw_database_mut()`, which grants access to they underlying salsa database to be able to write to it.

Because of the way we manage changes in the server, we have no need to initialise the file revision for a file in the `DidOpenTextDocument` handler, as VFS generates a `ChangeKind::Create` for it when we call `vfs.set_file_contents()` and it it the first time we see the file.

Empirically, we found this call would never return if we caused a lot of files to open while the initial background parsing step was going on.  Since it is unneccessary anyway, we remove it.

Result found by Roberto Aloi with input from Alan Zimmerman

Reviewed By: sopotc

Differential Revision: D51076342

fbshipit-source-id: 5ad0c638add214402f1bbf4dc46e1f81d566e2ec