HIR: put Body per Clause, not per Function
Summary:
This Diff is a preparation for updating the grammar to make separators (`;`, `.`) optional on function clauses, to improve the "parse broken code" experience. This is done by making a fun_decl only have one clause, and then stringing them back together again in HIR.

The preparation is to make `body::FunctionBody` not have its own `Body` or `BodySourceMap`, but to let each clause have its own. This is stored in `body::FunctionClauseBody`, and has most of the heavy lifting previously associated with the function itself.  This is more natural too, since each function clause is in fact independent of all the others for many kinds of analysis.

So the workhorse `InFunctionBody<T>` gets a companion `InFunctionClauseBody<T>`, which now does most of the work, and is used in the context of fold_function.

This diff, although large, is just a refactoring. The next diff will make the changes to the grammar, and reap the benefits.

Reviewed By: robertoaloi

Differential Revision: D51162080

fbshipit-source-id: 846fcfb53198619513015f90922d7078b783712f