grammar.js: no optional separator on function clause
Summary:
# Problem

With the current grammar, we expect a final `;` or `.` on a function clause, and will not return a parse for one if the terminator is missing.  Unfortunately tree-sitter is not able to automatically insert it, fundamentally because it is ambiguous which one to insert, without further semantic knowledge.

This leads to large syntax errors being reported, for the entire clause text, from the missing separator.

And of course newly-entered code does not yet have the separator.

# This Diff

To address this, we make the separator optional, and limit a function declaration to a single clause.  This does away with the extraneous syntax error, but leads to some complications inside ELP

- We now have one clause per fun_decl as parsed, so must reconstitute the full function declaration at the HIR level
- Any processing on the lowe level ast needs to be aware that an enclosing function will not have all the clauses
- Head mismatch and missing separator diagnostics become more complex

This diff builds on the work done in D51162080, which treats each function clause as a unit, having an HIR Body for each.  It keeps a `FunctionId` for an ast-level function declaration, which is now limited to a single clause.

It also introduces a `FunctionDefId` which refers to a re-constituted function, and indexes to a `FunctionDef`, and let the original `FunctionId` index to a `FunctionClauseDef`.

Insteady of using `InFunctionBody` as the workhorse, we use `InFunctionClauseBody`, and update the folds to use the `ClauseId` to retrieve the appropriate one from the function `Body`.

# To do

This diff does not deal with top-level macros that expand into (possibly multiple) function clauses.   That is dealt with in the next diff in this stack.

# Show and Tell

## Before

 {F1150769531}

## Using this diff

 {F1150769629}

Reviewed By: perehonchuk

Differential Revision: D51346431

fbshipit-source-id: c4ae27e1e959bce19d675f5551a5becec7432144