fix top-level macro clause lowering panic
Summary:
D51350855 brought in the ability to expand top level functions from macros, with the new grammar.js.
These are lowered as a recursive invocation of the body lowering machinery, with the macros in effect in the file preset.
When the body lowering finishes, there are some sanity checks. These should only be applied when the top-level body lowering finishes.

Reviewed By: TD5

Differential Revision: D51551386

fbshipit-source-id: d6cb0f18d8126429cf5e1dce50f5de7c2c411df4