Fix stale erlang_service diagnostics bug
Summary:
When we get new diagnostics we compare them with the old, and only update if there is a difference.
The erlang service diagnostics are wrapped in a `LabeledDiagnostics` struct so they can be allocated to syntax errors for undefined functions to avoid cascading errors.

The comparison logic checked that the LabeledDiagnostics::len() function for old and new were the same, then assumed it was safe to compare with a zip operation over iter() on the two.
Unfortunately the `len()` function does not return the number of diagnostics, but the number of labels, so the zip-based compare would cut off on the length of the shorter of the two lists.

The fix is to use `itertools::equal`, to avoid iterating twice, once to get the length, but still doing a full element-by-element equality check.

Reviewed By: TD5

Differential Revision: D51560268

fbshipit-source-id: cf821cbb4cd049167cfa5bebab646cf65a08f076