Only highlight LHS of a redundant match
Summary:
When one assigns an expression to an unused variable, the "trivial match" diagnostic (W0007) fires and the full expression gets underlined in yellow.

This leads to an unnecessary wall-of-yellow, that can be distracting. We should instead just highlight the lhs of the assignment. We do this in this diff.

Reviewed By: alanz

Differential Revision: D51563066

fbshipit-source-id: 91782858fcab6fbc1794e21fd77a854e9a80123c