server: move all changed file processing into one place
Summary:
With the watches we set up, we get a notification for every file that changes on disk, which includes the ones we save.
To ensure we do the same processing regardless of the change origin, refactor it in to a separate function called from both `DidChangeTextDocument` and `DidChangeWatchedFiles`.
This also keeps things robust for clients that do not support file watching.

Reviewed By: ir-regular

Differential Revision: D51548659

fbshipit-source-id: d824813e143e02581f7a8a831de9f3a344d5b871