BE: Start introducing fold::Strategy::VisibleMacros and InvisibleMacros
Summary:
This diff stack moves towards only using `fold::Strategy` to decide whether macros are visible or not.
In this step we introduce the new enum variants, `Strategy::VisibleMacros` and `Strategy::InvisibleMacros`.
We use the strategy to choose the appropriate `FoldBody` variant, and so start getting rid of `fold::WithMacros`.

Reviewed By: robertoaloi

Differential Revision: D51660226

fbshipit-source-id: 6f44d0de1b07df69b76b6ad7e8424e5071c56713