Abort eqWAlization earlier on error
Summary:
Previously, if an error occurred, then ELP would send a signal to eqWAlizer to abort, but would then still try to send an `ELPExitingModule` signal. eqWAlizer would not be able to receive it and the process would then crash.

This changes ELP's behaviour to stop before sending the `ELPExitingModule` signal if an error occurs.

Reviewed By: alanz

Differential Revision: D51706823

fbshipit-source-id: fb5673245e95253a6da3a6327d71a8fb099c0d94