def_map: when allocating function clauses to top level macros, consider prior separator
Summary:
A top-level function can have a separator state of `Dot`, `Semi`, or `Missing`, or be missing completely.
If we are allocationg a top-level macro expanding into a function clause, we check that the prior separator is explicitly `Semi` before combining them into the same top level function.

An alternative would be to combine unless there is a `Dot`.  This seems safer.

Reviewed By: perehonchuk

Differential Revision: D51708605

fbshipit-source-id: 345ecc461b7c8d77343c63c3365cd3b9b38d5c21