expand_macro for top level
Summary:
When we lower a function, we recursively call lowering on each clause, so each gets their own independent `Body`.
This lowering can only happen if we have the RHS of the macro to be expanded, which is in the macro stack of the original lowering process.  So transfer macro resolutions as well as the macro stack when recursing.

Also put the function clause name into FunctionClauseBody, because when it comes from a macro we only know this when we expand it.

Reviewed By: perehonchuk

Differential Revision: D51710752

fbshipit-source-id: 54d226fa72356282ea9596e173035a8075f2820d