Add codemod for onboarding suites to appatic
Summary:
# Context
We are rolling out a new CT hook to handle application lifetime and environments , essentially "auto_meckanic for applications". We want to run a similar codemod similar to the one for auto_meckanic to enable it on suites and remove all application handling logic that becomes redundant

# This diff
We introduce a new codemod, based on codemod_automeckanic. Right now the logic is:

- Add `{appatic, #{enable_autoclean => true}}` to suite/0 if the suite calls application or flexi functions
- Remove calls that stop applications or change their environment from end_per functions, as this is done managed by appatic
- Remove such calls from the end of testcases, allowing for trailing `ok`, etc.
- Remove calls to wa_test:stop_applications from end_per functions

# Coming later
Later in the stack we want to add support for:
- Simplifying end_per functions that by now are likely to become just "ok"
- Removing things like `[{apps, Apps} | Config]` from init_per functions, that will now be meaning less since `wa_test:stop_applications()` will be gone

Reviewed By: alanz

Differential Revision: D51988529

fbshipit-source-id: eb760510a28f5854cb9c12356fe0fa3ffe9bf26f